Rename stub files
