+++ __tests__/stubs/env-stubs.test.ts
@@ -1,6 +1,6 @@
 import { jest } from '@jest/globals'
-import { ResetCoreMetadata } from '../../src/stubs/core-stubs.js'
-import { EnvMeta, ResetEnvMetadata } from '../../src/stubs/env-stubs.js'
+import { ResetCoreMetadata } from '../../src/stubs/core.js'
+import { EnvMeta, ResetEnvMetadata } from '../../src/stubs/env.js'
 import type { EnvMetadata } from '../../src/types.js'
 
 /** Empty EnvMetadata Object */
