+++ src/commands/run.ts
@@ -1,8 +1,8 @@
 import { config } from 'dotenv'
 import { createRequire } from 'module'
 import quibble from 'quibble'
-import { CORE_STUBS, CoreMeta } from '../stubs/core-stubs.js'
-import { EnvMeta } from '../stubs/env-stubs.js'
+import { CORE_STUBS, CoreMeta } from '../stubs/core.js'
+import { EnvMeta } from '../stubs/env.js'
 import type { Action } from '../types.js'
 import { printTitle } from '../utils/output.js'
 import { isESM } from '../utils/package.js'
